started getting all to work together

+++ Projectiles/yeahyeqah.py
@@ -7,11 +7,11 @@ def run_sim():
     pg.init()
     with open('PVars.txt', "r") as file:
         # Read each line and assign values to variables
-        user_angle = int(file.readline().strip())
-        user_mass = int(file.readline().strip())
-        user_initial_vel = int(file.readline().strip())
-        user_starting_height = int(file.readline().strip())
-        user_end_time = int(file.readline().strip())
+        user_angle = float(file.readline().strip())
+        user_mass = float(file.readline().strip())
+        user_initial_vel = float(file.readline().strip())
+        user_starting_height = float(file.readline().strip())
+        user_end_time = float(file.readline().strip())
     # Simulation setup
     mass = user_mass
     angle = user_angle  # Initial angle in degrees
